Move to uv from pdm.
It's much faster, and pdm was failing to upgrade desktop-notifier for some reason.
Though after uv was able to install the latest desktop-notifier, the code now doesn't
pass through pyright, so pinning it on the old version for now.

Meanwhile Textual also got upgraded, and broke some of the tests.  Pinning it on an
older version as well.